reviews/pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
  "mypy<1.0,>=0.971",
  "flake8<6.0.0,>=5.0.3",
  "black>=22.6.0",
]

[tool.pdm.build]
includes = []

[tool.pdm.scripts]
server = "uvicorn app:app --port 4002 --reload"

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=1.0.0"]

[tool.mypy]
plugins = [
  "strawberry.ext.mypy_plugin",
]

[project]
authors = [
  {name = "Patrick Arminio", email = "patrick.arminio@gmail.com"},
]
dependencies = ["strawberry-graphql==0.124.0.dev.1659525319", "starlette<1.0.0,>=0.16.0", "uvicorn<1.0.0,>=0.13.4", "taskipy<2.0.0,>=1.8.1"]
description = ""
license = {text = "MIT"}
name = "reviews"
requires-python = ">=3.9,<4.0"
version = "0.1.0"

[project.urls]

[project.optional-dependencies]
